get initial memory, move cprofile decorator
